add optimizer_context
